requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.2
contourpy==1.1.0
cycler==0.11.0
einops==0.6.1
filelock==3.12.2
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.4.2
kiwisolver==1.4.5
lightning-utilities==0.9.0
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
MedPy==0.4.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nibabel==3.2.2
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.3
python-dateutil==2.8.2
pytorch-lightning==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==1.3.2
scipy==1.10.1
SimpleITK==2.2.1
six==1.16.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2023.7.10
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.1.0
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
